% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_diversity.R
\name{alpha_diversity}
\alias{alpha_diversity}
\title{Measure Composition of a Full Image}
\usage{
alpha_diversity(img, img_type, col_pal, plot_bkgd = "W")
}
\arguments{
\item{img}{Named List: 4d array of the image whose composition is to be measured}

\item{img_type}{Character: whether the image represents "O"bjects or "S"calars}

\item{col_pal}{Named List: color palette for the image}

\item{plot_bkgd}{Character: whether the plot background should be "W"hite or "B"lack}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data Frame: Percentage of image belonging to each variable and a plot of this data
}\if{html}{\out{</div>}}
}
\description{
Quantify the object or scalar composition of an entire image
}
\details{
Quantify and plot the fraction of an image that belongs to each object or that is positive for each scalar.
The number of different object or scalars and how evenly represented they are throughout the image
determines the image's alpha diversity. Alpha diversity is not available for object-specific scalars.
}
