% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_diversity.R
\name{beta_diversity}
\alias{beta_diversity}
\title{Measure Compositions of Objects Separately}
\usage{
beta_diversity(img, img_type, col_pal, plot_bkgd = "W")
}
\arguments{
\item{img}{Named List: 4d arrays of the images of the parent objects and the constituent objects or scalars}

\item{img_type}{Character: whether the constituent variables are "O"bjects or "S"calars}

\item{col_pal}{Named List: color palettes for the object images and the constituent objects or scalars}

\item{plot_bkgd}{Character: whether the plot background should be "W"hite or "B"lack}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame: object or scalar composition of each parent object separately and a plot of this data
}\if{html}{\out{</div>}}
}
\description{
Quantify the object or scalar composition of each parent object in an image separately
}
\details{
Quantify and plot the fraction of each parent object that belongs to each sub-object or that is positive for each
scalar. The difference in composition across the parent objects determines the image's beta diversity.
The number of different objects or scalars and how evenly represented they are throughout each parent object
determines the parent object's alpha diversity, which is also reported.
}
