% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_distributions.R
\name{build_dist}
\alias{build_dist}
\title{Build a Joint Distribution}
\usage{
build_dist(census, vars, focal_vars = "all")
}
\arguments{
\item{census}{Data frame: rounded percentages of each variable across neighborhoods}

\item{vars}{Vector: variables for which to build the distribution}

\item{focal_vars}{Vector: all variables in a broader analysis}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame: all possible quantitative combinations of the variables and their counts
}\if{html}{\out{</div>}}
}
\description{
Assemble a joint distribution for an ensemble from a rounded census.
}
\details{
After the variables in a census have been rounded into bins, assemble a joint distribution for an ensemble.
}
