% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_vol}
\alias{calc_vol}
\title{Calculate Volume of a Neighborhood of a Given Size}
\usage{
calc_vol(L, dims)
}
\arguments{
\item{L}{Named Vector: radius of neighborhood in pixels in X, Y, and Z}

\item{dims}{Vector: dimensions of overall image in pixels in X, Y, and Z}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Vector: volumes of neighborhoods in pixels
}\if{html}{\out{</div>}}
}
\description{
Calculate the volume in pixels for a single neighborhood radius
}
