% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_vols}
\alias{calc_vols}
\title{Calculate Volumes of Neighborhoods of Given Sizes}
\usage{
calc_vols(L, dims)
}
\arguments{
\item{L}{List: named vectors of radii of neighborhoods in pixels in X, Y, Z}

\item{dims}{Vector: dimensions of overall image in pixels in X, Y, and Z}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Numeric: area of neighborhood in pixels
}\if{html}{\out{</div>}}
}
\description{
Calculate volumes in pixels for different neighborhood radii
}
