% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_image_table.R
\name{census_image}
\alias{census_image}
\title{Census Images}
\usage{
census_image(
  images,
  OS_pairs = NULL,
  seed_points = NULL,
  radii,
  sample_size,
  cores = NULL
)
}
\arguments{
\item{images}{Named List: 4d array/s representing object maps and/or scalar sets}

\item{OS_pairs}{Named List: link table/s that specify object-scalar pairs of interest}

\item{seed_points}{Named List: vectors specifying which objects or scalars may be seed points. NULL for all is default; NA for none.}

\item{radii}{Named List: vectors of X, Y, and Z pixel radii at each micron radius}

\item{sample_size}{Numeric Vector: number of neighborhoods to draw at each radius}

\item{cores}{Integer: number of cores to use for parallel calculations}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ List: data frame of the census and nested list of the patch list
}\if{html}{\out{</div>}}
}
\description{
Census neighborhoods across object maps and/or scalar sets from a single biological speciment
}
\details{
Draw neighborhoods on a list of images representing a single biological specimen. All images must
have the same dimensions. All images must be named as "O"bject or "S"calar, followed by a numeric code. If both
object maps and scalar sets are included, each scalar set may be independent or linked.
In the latter case, the object-scalar pairs of interest must be given with a link table
with the same name as the involved object map. For each image, seed points control
where neighborhoods may be centered. By default, seed points are wherever any object or scalar exists.
The radii and number of neighborhoods to draw at each radius should also be specified.
}
