% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_image_table.R
\name{census_patches}
\alias{census_patches}
\title{Census the Patches in a Neighborhood}
\usage{
census_patches(sub_lyr, incl_pix, image_type, bin_thr)
}
\arguments{
\item{sub_lyr}{4d Numeric Array: bounding box of one neighborhood, excised from a larger image across all slices}

\item{incl_pix}{3d Logical Array: mask of pixels included in a spherical neighborhood within a bounding box}

\item{image_type}{Character Vector: Object map or scalar set ID of each slice of the image}

\item{bin_thr}{Numeric Vector: binary threshold for each slice of the image}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame: scalar intensity or object ID and pixel count for each patch
}\if{html}{\out{</div>}}
}
\description{
Record sizes of contiguous patches of pixels in a neighborhood, plus their object IDs and/or scalar intensities
}
\details{
Find contiguous patches of pixels that belong to the same object or that are positive for the same scalar.
When scalars are linked to objects, the total scalar intensity on each object patch is recorded, as well.
}
