% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_image_table.R
\name{census_table}
\alias{census_table}
\title{Census a Table}
\usage{
census_table(
  object_table,
  OS_pairs = NULL,
  seed_points = NULL,
  radii,
  sample_size,
  cores = NULL
)
}
\arguments{
\item{object_table}{Data Frame: centroid coordinates, object type, and mean expression of scalars}

\item{OS_pairs}{Named List: link table that specifies object-scalar pairs of interest}

\item{seed_points}{Named List: vector specifying which objects may be seed points, or NULL for no restrictions}

\item{radii}{Vector: radii of neighborhoods, in the same units as the centroid coordinates, typically microns}

\item{sample_size}{Vector: number of neighborhoods to draw at each radius, no more than the number of objects}

\item{cores}{Integer: number of cores to use for parallel calculations}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ List: data frame of the census and nested list of the patch list
}\if{html}{\out{</div>}}
}
\description{
Census neighborhoods from a table of objects
}
\details{
Draw neighborhoods on an object table to record the contents of each. The object table must give the
centroid coordinates in columns labeled "X", "Y", and "Z", and the numeric code for object ID in a column labeled
"Object". Optional further columns may give the intensity of scalars. If scalar intensities are included, the
object-scalar pairs of interest must be specified in a link table with the same name as the involved object map.
Seed points control where neighborhoods may be centered - by default, wherever any object exists.
The radii and number of neighborhoods should also be specified.
}
