% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{entropy}
\alias{entropy}
\title{Calculate the Histogram Entropy Estimate of a Distribution}
\usage{
entropy(cen, bin)
}
\arguments{
\item{cen}{Array 2d: matrix of neighborhoods as rows and variables as columns, giving bin ID for each entry}

\item{bin}{Integer: number of bins per variable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Numeric: histogram entropy estimate for the ensemble
}\if{html}{\out{</div>}}
}
\description{
Calculate entropy or joint entropy given for an ensemble of variables
}
