% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_map_patterns.R
\name{learn_pattern}
\alias{learn_pattern}
\title{Learn a Spatial Pattern}
\usage{
learn_pattern(
  census,
  ensemble,
  radius,
  col_pal,
  group = NULL,
  focal = NULL,
  reference = NULL,
  patch_list = NULL,
  conf_int = 0.95,
  obj_norm = NULL,
  scl_norm = "all",
  som_reps = 50,
  toroidal = F,
  smooth_window = 100,
  sub_sample = F,
  plot_bkgd = "W"
)
}
\arguments{
\item{census}{Data frame: census for a single image, or list of censuses for multiple images}

\item{ensemble}{Vector: variables included in the ensemble of interest}

\item{radius}{Numeric: radius of neighborhoods to use}

\item{col_pal}{List: named list of the color palettes encompassing all variables in the ensemble}

\item{group}{Frame: data frame with grouping factors as columns and images as rows}

\item{focal}{Character: group/s for which to learn the covariation pattern}

\item{reference}{Character: group/s against which to compare the covariation pattern}

\item{patch_list}{Nested List: patch list must be included if comparisons are made vs. random}

\item{conf_int}{Numeric: confidence interval of local representation to display along latent path}

\item{obj_norm}{Character: how to normalize local abundance of objects: NULL for none, "ind" for individually, "all" for collectively}

\item{scl_norm}{Character: how to normalize local abundance of scalars: NULL for none, "ind" for individually, "all" for collectively}

\item{som_reps}{Numeric: number of learning repetitions to run the self-organizing map}

\item{toroidal}{Logical: whether the self-organizing map should be toroidal or not}

\item{smooth_window}{Numeric: number of neighborhoods to smooth together}

\item{sub_sample}{Logical: whether to subsample census neighborhoods for speed}

\item{plot_bkgd}{Character: whether to plot on a "W"hite or "B"lack background}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame: local content of each variable along the latent path, also returned graphically
}\if{html}{\out{</div>}}
}
\description{
Learn the spatial patterning of an ensemble by drawing a self-organized map on its joint distribution
}
\details{
Draw a curve through the regions of highest data density on an ensemble's joint distribiution.
All neighborhoods in the census are reordered to follow the trajectory of this curve.
This reveals how the variables in the ensemble covary across the entire spatial extent.
Enrichment of features is plotted versus a reference - randomized expectations if a patch list is given,
or a different group of specimens if focal and reference groups are defined..
}
