% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image_table.R
\name{load_image}
\alias{load_image}
\title{Load an Image}
\usage{
load_image(in_file, img_type, bkgd_col, num_chs, keep_chs = NULL)
}
\arguments{
\item{in_file}{Character: path to the input file}

\item{img_type}{Character: whether the image is a "O"bject or "S"calar image}

\item{bkgd_col}{Character: for object images, the background color's hex code/nickname, or NULL if none}

\item{num_chs}{Numeric: for scalar images, the number of channels in the image}

\item{keep_chs}{Vector: for scalar images, the channels to keep upon loading, or NULL if all}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ List: 4d array representing the image: x,y,z,c. And for object maps, the extracted color palette
}\if{html}{\out{</div>}}
}
\description{
Load an object or scalar image and transform it into a 4d array
}
\details{
Load an image into array format. Object images must be RGB color .tif. Scalar images must be 8-bit grayscale .tif
possibly including multiple channels. 3d or 2d images are permitted.
}
