% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image_table.R
\name{load_table}
\alias{load_table}
\title{Load a Table}
\usage{
load_table(in_file, table_type, img, col_pal)
}
\arguments{
\item{in_file}{Character: path to the table file}

\item{table_type}{Character: whether the input is a "P"rofile, "O"bject, or "L"ink table}

\item{img}{Array 4d: for profile tables, the object image to which the table corresponds}

\item{col_pal}{Vector: for profile tables, the color palette to which the table corresponds}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame: the profile table where objects are reordered to match the corresponding object image, or the object or link table exactly as in the source file
}\if{html}{\out{</div>}}
}
\description{
Load a profile, object, or link table
}
\details{
Load a table in .csv format. The table can be a profile, object, or link table.
For profile tables, column 1 must be labeled "Object" and give numeric object IDs.
Column 2 must be labeled "Count" and give the count of pixels or items belonging to each object.
Optionally, columns 3-5 can labeled "X", "Y", and "Z" and give the coordinates of one pixel belonging
to each object in the image, to speed the matching of objects between the table and the corresponding image.
The remaining columns give the compositional profile of each object, in terms of scalars or sub-objects.
For object tables, columns 1-3 must be labeled "X", "Y", and "Z" and give the centroid coordinates for each object.
Column 4 must be labeled "Object" and give the numeric code for each object ID. Optional further columns
may give scalar intensity values. For link tables, the rows must be labeled with the alpha-numeric codes of the
objects, e.g. "O1.1", and the columns must be labeled with the alpha-numeric codes of the scalars, e.g. "S1.1".
The object set and scalar set must share the same numeric code. Each entry is 1 or 0, indicating whether the
indicated combination should be analyzed or ignored.
}
