% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image_palette.R
\name{make_palette}
\alias{make_palette}
\title{Make a Color Palette}
\usage{
make_palette(num_cols)
}
\arguments{
\item{num_cols}{Numeric: number of colors}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Vector: hex codes for each color in the palette
}\if{html}{\out{</div>}}
}
\description{
Randomly select contrasting colors to form a palette of a specified size
}
\details{
Generate a set of color with maximal contrast. Beyond 12 colors, visual distinction can still be difficult.
}
