% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_map_patterns.R
\name{map_pattern}
\alias{map_pattern}
\title{Map a Pattern back to an Image}
\usage{
map_pattern(
  covar_data,
  region_bounds,
  img,
  census,
  radius,
  radii,
  col_pal,
  orig_bkgd = T
)
}
\arguments{
\item{covar_data}{Data frame: covariation plot data}

\item{region_bounds}{List: non-overlapping 2-entry vectors of gates on the covariation plot}

\item{img}{Named List: 4d arrays of the image/s on which the covariation data is based}

\item{census}{Data frame: census of the image/s from which the covariation plot was made}

\item{radius}{Numeric: radius at which the covariation plot was generated}

\item{radii}{Named List: for each radius in microns, named vectors of the appropriate radii in pixels in X, Y, and Z}

\item{col_pal}{Vector: color palette for newly defined objects}

\item{orig_bkgd}{Logical: when mapping, whether to impose the background from the original image}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ List: new object image in array format and its corresponding color palette
}\if{html}{\out{</div>}}
}
\description{
Use manually-defined gates on a covariation plot to generate an image of custom-defined objects
}
\details{
Map features of a covariation plot onto the original image. The covariation plot is partitioned into
regions, and the census is used to determine which region each pixel in the image belongs to. New colors are
assigned to these regions, resulting in a new object image.
}
