% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_mutual_info.R
\name{measure_cisMI}
\alias{measure_cisMI}
\title{Calculate Cis Mutual Information}
\usage{
measure_cisMI(
  census,
  patch_list,
  depth,
  radii = NULL,
  bootstraps = 100,
  all = NULL,
  alo = NULL,
  not = NULL,
  max_bins = 100,
  cores = NULL
)
}
\arguments{
\item{census}{Data frame: census from the biological specimen}

\item{patch_list}{List: patch list from the biological specimen}

\item{depth}{Numeric: maximum number of variables per ensemble}

\item{radii}{Vector: radii to examine, defaulting to all radii present in the provided census}

\item{bootstraps}{Numeric: number of randomized censuses to generate for estimating confidence intervals}

\item{all}{Vector: variables which all must be included in every maximum-depth ensemble}

\item{alo}{Vector: variables of which at least one must be included in every maximum-depth ensemble}

\item{not}{Vector: Variables which should not be included in any maximum-depth ensemble}

\item{max_bins}{Numeric: maximum number of bins per variable for rounding}

\item{cores}{Numeric: number of cores to use for parallel calculations}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  List: data frames of cisMI for all ensembles at each radius
}\if{html}{\out{</div>}}
}
\description{
Calculate cis mutual information for many ensembles and radii from a census
}
\details{
Quantify the non-randomness of patterning for each ensemble that is not explained by patterning among any
subset of those variables, for a single biological specimen. This is statistically compared to null expectations
in the absence of patterning, via simulation of randomized censuses.
}
