% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_mutual_info.R
\name{measure_transMI}
\alias{measure_transMI}
\title{Calculate Trans Mutual Information}
\usage{
measure_transMI(
  censuses,
  groups,
  depth,
  radii,
  bootstraps = 100,
  all = NULL,
  alo = NULL,
  not = NULL,
  max_bins = 100,
  cores = NULL
)
}
\arguments{
\item{censuses}{List: censuses for multiple biological specimens}

\item{groups}{Frame: data frame with grouping factors as columns and specimens as rows}

\item{depth}{Numeric: maximum number of variables per ensemble}

\item{radii}{Numeric: radii at which to examine spatial patterns}

\item{bootstraps}{Numeric: number of randomized transMI networks to simulate for estimating P values}

\item{all}{Vector: variables which all must be included in every maximum-depth ensemble}

\item{alo}{Vector: variables of which at least one must be included in every maximum-depth ensemble}

\item{not}{Vector: Variables which should not be included in any maximum-depth ensemble}

\item{max_bins}{Numeric: maximum number of bins per variable for rounding}

\item{cores}{Numeric: number of cores to use for parallel calculations}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  List: data frames of transMI for all ensembles at each radius
}\if{html}{\out{</div>}}
}
\description{
Calculate trans mutual information for every ensemble given a grouping structure of specimens
}
\details{
Quantify the difference in patterning for every ensemble, across groups of multiple biological specimens.
This is statistically compared to null expectations in which sample groups are defined at random.
}
