% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merge_table.R
\name{merge_objects}
\alias{merge_objects}
\title{Merge Objects Together}
\usage{
merge_objects(prof_table, img, col_pal, obj_table, obj_groups)
}
\arguments{
\item{prof_table}{Data frame: profile table containing the objects to be merged}

\item{img}{3d Array: image containing the objects to be merged}

\item{col_pal}{Vector: Color palette for the objects including those to be merged}

\item{obj_table}{Data frame: object table including the objects to be merged}

\item{obj_groups}{List: vectors of numeric object IDs, grouping together the objects to be merged}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ List: profile table, object image, color palette, and/or object table, but with objects merged
}\if{html}{\out{</div>}}
}
\description{
Merge objects together in a profile table, image, color palette, and/or object table
}
\details{
Merge similar objects together in a profile table, image, color palette, and/or object table that all correspond.
}
