% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition_evenly}
\alias{partition_evenly}
\title{Partition a Number into Addends as Evenly as Possible}
\usage{
partition_evenly(total, comps)
}
\arguments{
\item{total}{Numeric: whole number to be partitioned into additive components}

\item{comps}{Numeric: number of additive components to split the total into}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Vector: addends into which the total decomposes
}\if{html}{\out{</div>}}
}
\description{
Partition a number of variables into addends as evenly as possible
}
