% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{patch_3D}
\alias{patch_3D}
\title{Find Connected Patches in a 3D Mask}
\usage{
patch_3D(mask)
}
\arguments{
\item{mask}{Array 3d: logical array of pixels of a particular object or positive for a particular scalar}
}
\value{
list of the 3d array of patch indices and the 3d array of patch sizes
}
\description{
Find connected patches and their sizes in a 3d mask. Based on buchsbaum's neuroim2::connComp3D
}
