% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot a Distribution}
\usage{
plot_dist(
  census,
  ensemble,
  radius,
  bin_num = NULL,
  patch_list = NULL,
  plot_zoom = F,
  plot_bkgd = "W"
)
}
\arguments{
\item{census}{Data frame: census of the specimen}

\item{ensemble}{Vector: 1 or 2 variables for which to plot the distribution}

\item{radius}{Numeric: radius for which to plot the distribution}

\item{bin_num}{Numeric: number of rounding bins per variable; if omitted, a scatter plot is given}

\item{patch_list}{List: if randomizing, patch list for the corresponding image or table}

\item{plot_zoom}{Logical: whether to zoom into the distribution, or keep a 0-100\% range for all variables}

\item{plot_bkgd}{Character: whether the plot background should be "B"lack or "W"hite}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame: the joint distribution of the ensemble, and a plot
}\if{html}{\out{</div>}}
}
\description{
Plot how often each quantitative combination of 1 or 2 variables occur in a specimen
}
\details{
Plot a distribution of co/occurrence for one or two variables. Omitting a number of bins for rounding to obtain
a 1d density or 2d scatter plot, reflecting how the data exists in the census. Provide a number of bins for
rounding to obtain a probability distribution across bins, reflecting how the data is used for mutual information.
Omit a patch list to obtain the plot for the observed census. Provide a patch list to obtain the plot for a
randomized census. More than two variables and a patch list can be provided if only a randomized census is desired,
and no plot will be returned.
}
