% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image_palette.R
\name{plot_image}
\alias{plot_image}
\title{Plot an Image}
\usage{
plot_image(
  img,
  img_type,
  col_pal,
  slice = c(Z = 1),
  objects = NULL,
  scalars = NULL,
  enh_cnt = 0
)
}
\arguments{
\item{img}{Array 4d: the image to plot}

\item{img_type}{Character: whether the image represents "O"bjects or "S"calars}

\item{col_pal}{Vector: color palette for the objects or scalars}

\item{slice}{Named Vector: "X", "Y", or "Z" coordinate at which to slice}

\item{objects}{Vector: for object images, numeric codes of the objects to visualize, or all of them with NULL}

\item{scalars}{Vector: for scalar images, numeric codes of the scalars to visualize, or all of them with NULL}

\item{enh_cnt}{Numeric: quantile of pixels to saturate for each scalar}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Array 4d: array representation of the image that was plotted.
}\if{html}{\out{</div>}}
}
\description{
Plot an image from its array representation
}
\details{
Visualize either the entirety of a 2d image or a slice of a 3d image. Subsets of the image's scalars or objects
may be specified to reduce what is plotted. The plotted image is also saved as a .tif file.
}
