% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutual_info.R
\name{plot_MI_radius}
\alias{plot_MI_radius}
\title{Plot Mutual Information by Radius}
\usage{
plot_MI_radius(
  mi,
  ensemble,
  p_thr = NULL,
  p_adj = F,
  group = NULL,
  plot_bkgd = "W"
)
}
\arguments{
\item{mi}{List: cis or trans MI measurements across all radii and ensembles}

\item{ensemble}{Numeric: vector of the variables included in the ensemble of interest}

\item{p_thr}{Numeric: threshold corrected p-value, defaulting to 0.05}

\item{p_adj}{Logical: whether to use multiple-testing-corrected p values or raw p values}

\item{group}{Character: for transMI, name of the grouping factor to consider if there are more than 1}

\item{plot_bkgd}{Character: whether the plot background should be "B"lack or "W"hite}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data Frame: the focal ensemble's cis/trans MI scores and a plot of this data
}\if{html}{\out{</div>}}
}
\description{
Plot cisMI or transMI scores across radii, for a single ensemble
}
\details{
Compare across length scales how non-randomly one ensemble is patterned, in the case of cisMI,
or how well its patterning distinguishes specimen groups, in the case of transMI.
}
