% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutual_info.R
\name{plot_MI_rank}
\alias{plot_MI_rank}
\title{Plot Mutual Information by Rank}
\usage{
plot_MI_rank(
  mi,
  radius,
  depth = NULL,
  col_pals = NULL,
  p_thr = NULL,
  mi_thr = NULL,
  p_adj = T,
  all = NULL,
  alo = NULL,
  not = NULL,
  group = NULL,
  plot_bkgd = "W"
)
}
\arguments{
\item{mi}{List: cis or trans MI measurements across all radii and ensembles}

\item{radius}{Numeric: radius for which to plot P values}

\item{depth}{Vector: ensemble depths to include in the plot}

\item{col_pals}{List: list of the palettes for each object map or scalar set represented in the ensembles}

\item{p_thr}{Numeric: threshold corrected p value to consider, defaulting to 0.05}

\item{mi_thr}{Numeric: threshold MI value to consider, defaulting to 0.1 for cisMI or 0 for transMI}

\item{p_adj}{Logical: whether to use multiple-testing-corrected p values or raw p values}

\item{all}{Vector: variables which all must be included in every ensemble of the highest depth}

\item{alo}{Vector: variables of which at least one must be included in every ensemble of the highest depth}

\item{not}{Vector: variables which should not be included in any ensemble}

\item{group}{Character: for transMI, name of the grouping factor to plot}

\item{plot_bkgd}{Character: whether the plot background should be "B"lack or "W"hite}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data Frame: all significant ensembles and their cis/trans MI scores and a plot of this data
}\if{html}{\out{</div>}}
}
\description{
Plot cisMI or transMI scores across ensembles, sorted from most to least significant
}
\details{
Compare ensembles by how non-randomly they are patterned, in the case of cisMI, or by
how well their patterning distinguishes specimen groups, in the case of transMI.
}
