% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image_palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot a Color Palette}
\usage{
plot_palette(col_pal, axis_label, col_labels, vertical = F, plot_bkgd = "W")
}
\arguments{
\item{col_pal}{Vector: color palette to plot}

\item{axis_label}{Character: label for the entire palette of colors}

\item{col_labels}{Vector: labels for individual colors}

\item{vertical}{Logical: whether to plot the colors vertically or horizontally}

\item{plot_bkgd}{Character: whether the plot background should be "W"hite or "B"lack}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Nothing: only a plot of the color palette is created
}\if{html}{\out{</div>}}
}
\description{
Plot all colors in a palette with labels
}
\details{
Plot a color palette, optionally including short names for what each  variable represents biologically.
}
