% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merge_table.R
\name{plot_table}
\alias{plot_table}
\title{Plot a Profile Table}
\usage{
plot_table(
  prof_table,
  compare = "A",
  normalize = "U",
  tile_plots = F,
  plot_bkgd = "W"
)
}
\arguments{
\item{prof_table}{Data frame: profile table with object IDs, pixel/object count, and composition of each object}

\item{compare}{Character: whether to normalize "A"cross objects, "W"ithin objects, or "B"oth}

\item{normalize}{Character: whether to normalize "U"niformly or by "Z" score}

\item{tile_plots}{Logical: whether plots should be tile or separate}

\item{plot_bkgd}{Character: whether the plot should have a "W"hite or "B"lack background}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Nothing: a plot of the heatmap is generated
}\if{html}{\out{</div>}}
}
\description{
Plot a heatmap to show the compositions of each object in terms of scalars
}
\details{
Plot a heatmap of object compositional profiles from a profile table. Column 1 must be labeled "Object" and give
numeric object IDs. Column 2 must be labeled "Count" and give the count of pixels or items belonging to each object.
The remaining columns give the compositional profile of scalar expression for each object.
Normalizing across objects is recommended: scalars are relative to their individual min vs. max levels across all objects.
Normalizing within objects is also possible: scalars are relative to their collective min vs. max within individual objects.
Normalizing uniformly is recommended, but normalizing by Z-score is also possible.
Bar plots show the number of pixels or items belonging to each object, and the maximum intensity for each scalar,
i.e. what quantitative value bright yellow indicates. For the latter plot, higher bars are better.
Low bars indicate that the scalar has very low and/or sparse intensity, even among the highest expressing objects.
}
