% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_column}
\alias{round_column}
\title{Round Values of a Variable}
\usage{
round_column(focal_col, col_min, col_max, bin_num, bin_val)
}
\arguments{
\item{focal_col}{Numeric: a specific column of a census to round}

\item{col_min}{Numeric: minimum value to round to}

\item{col_max}{Numeric: maximum value to round to}

\item{bin_num}{Numeric: number of bins to round into}

\item{bin_val}{Logical: whether to include bin values or just give them integer IDs}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Vector: the rounded values of the census column
}\if{html}{\out{</div>}}
}
\description{
Round the values of a variable found in a column of a census
}
