% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_distributions.R
\name{smooth_dist}
\alias{smooth_dist}
\title{Regularize a Joint Distribution}
\usage{
smooth_dist(joint_dist, bin_num, min_max, full_dist)
}
\arguments{
\item{joint_dist}{Data frame: all possible quantitative combinations of variables and their counts}

\item{bin_num}{Numeric: number of rounding bins per variable; can be a vector of different numbers for each variable}

\item{min_max}{Data frame: minimum and maximum value for each variable in the joint distribution}

\item{full_dist}{Logical: whether to return the full distribution or just the vector of frequencies}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame: regularized frequencies of all possible quantitative combinations of the variables
}\if{html}{\out{</div>}}
}
\description{
Regularize a joint distribution to insure that all possible bins have positive probability
}
\details{
Regularize a joint distribution to assign small probabilities to all bins, even those with measured value 0.
}
