% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_image_table.R
\name{standardize_censuses}
\alias{standardize_censuses}
\title{Standardize Columns across Censuses}
\usage{
standardize_censuses(censuses)
}
\arguments{
\item{censuses}{List: Census data frames for every image in a multi-image data set}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ List: Census data frames for every image in a multi-image data set, with standardized columns
}\if{html}{\out{</div>}}
}
\description{
Enforce that all censuses contain the same columns in the same order
}
\details{
For a list of censuses meant to be analyzed together, enforce that censuses contain the same columns
in the same order.
}
