% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_parameters.R
\name{suggest_number}
\alias{suggest_number}
\title{Suggest Number of Neighborhoods to Census}
\usage{
suggest_number(coverage, radii, images)
}
\arguments{
\item{coverage}{Vector: desired image coverage/s}

\item{radii}{List: named vectors of neighborhood radii for the X, Y, and Z dimensions in pixels}

\item{images}{List: 3d arrays of all images to be analyzed}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Vector: Number of neighborhoods to census at each radius
}\if{html}{\out{</div>}}
}
\description{
Suggest the number of neighborhoods of each radius to census, based on a desired coverage
}
\details{
Calculate the number of neighborhoods to census, based on neighborhood radii and the desired coverage.
Coverage is the number of neighborhoods each non-background pixel is included in, on average.
Coverage should not exceed 5x, to prevent pseudo-replication.
}
