% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_parameters.R
\name{suggest_radii}
\alias{suggest_radii}
\title{Suggest Radii of Neighborhoods to Census}
\usage{
suggest_radii(target, pix_res)
}
\arguments{
\item{target}{Vector: desired radii in microns}

\item{pix_res}{Named Vector: resolution of the image in microns/pixel in the X, Y, and Z dimensions}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Vector: pixel radii to use for censusing
}\if{html}{\out{</div>}}
}
\description{
Suggest radii in pixels of neighborhoods to collect for the census, based on desired radii in microns
}
\details{
Calculate the neighborhood radii in pixels based on desired radii in microns and the image resolution
in X, Y, and Z.
}
