% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_image_table.R
\name{summarize_patches}
\alias{summarize_patches}
\title{Summarize a Patch List as a Census
From a patch list of the patches of each object and scalar inside each neighborhood of a specific radius,
create a census. Each row is a neighborhood and each column is a variable of interest, including
object abundance, avg scalar intensity at large, and avg scalar intensity attached to certain objects.}
\usage{
summarize_patches(nbhd_patches, OS_pairs)
}
\arguments{
\item{nbhd_patches}{List: Across all object maps and/or scalar sets, the contiguous patches in each neighborhood}

\item{OS_pairs}{List: matrices that specify object-scalar pairs of interest}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame: census of the objects and scalars in each neighborhood
}\if{html}{\out{</div>}}
}
\description{
Transform a patch list into a census
}
