% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{total_comp_bins}
\alias{total_comp_bins}
\title{Find Maximum Total Number of Compositional Bins}
\usage{
total_comp_bins(dimension, bins_per_var, min_per_var, max_per_var)
}
\arguments{
\item{dimension}{Numeric: total number of compositional variables in the joint distribution}

\item{bins_per_var}{Numeric: number of bins to split each compositional variable into}

\item{min_per_var}{Vector: minimum value of each compositional variable}

\item{max_per_var}{Vector: maximum value of each compositional variable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Numeric: the maximum number of possible compositional bins
}\if{html}{\out{</div>}}
}
\description{
Find the total number of possible joint bins for a joint distribution of compositional variables
}
