% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unwhich2d}
\alias{unwhich2d}
\title{Create a Logical Array}
\usage{
unwhich2d(xy_coors, matrix_dims)
}
\arguments{
\item{xy_coors}{Matrix: xy coordinates of points in an array}

\item{matrix_dims}{Vector: dimensions of a matrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Array: an array of the specified dimensions where the specified coordinates are T and all else F
}\if{html}{\out{</div>}}
}
\description{
From a list of True coordinates, create a logical array
}
